Implement sqlite-based completion.
Implements completion.models.sql, a sqlite based completion model.
This implements the base model and unit-tests for basic functionality,
but does not switch any of the existing completion models to use the
sqlite backend.

The API is designed to stray as little as possible from the
BaseCompletionModel API, but once all models are converted to the sqlite
backend it should be possible to simplify the API.