Use an in-memory db for sqlite.
This should speed up operations, and we don't need to store that much
data so space shouldn't be a problem (we had all the completions in
memory before anyways.

As an added bonus, we no longer need to drop tables before creating
them because we get a fresh database for each instance. We also no
longer need to mock out the cache dir for tests.

See https://sqlite.org/inmemorydb.html