Don't double-init url completion.
If you open URL completion shortly after opening qutebrowser, it will
init once to show the completion and once again shortly after due to
the signal emmitted by the async history read.

This means we are doing extra unnecessary work and the initial results
of history completion may be incomplete. With SQL, it causes the
additional problem that we will cause an error trying to create the
same table again. We could use `CREATE TABLE IF NOT EXISTS` but that
could mask other errors, and we would end up creating duplicate rows
anyways.

With this patch, the signal no longer triggers a url completion update.
Instead, init waits for the signal before it builds the url completion
model.