Simplify TabCompletionModel.rebuild.
Rather than trying to add/remove rows from the model to accomodate
new/removed tabs or windows, just clear the model and reconstruct it
from scratch.

The cost to do this is trivial and it simplifies the implementation
significantly. This will make it easier to port to a new completion API
as described in #74.