Adds ability to map config options to backend setting attributes.
Hoists config-to-attribute MAPPINGS from subclasses at init time and adds
get_attribute function to allow querying it.

All this is done at the module level because the instances of whatever
subclass is running isn't stored anywhere. And there is only going to be one
backend running at one time anyway.