Also emit urlChanged from acceptNavigationRequest.
Emitting urlChanged this early in the lead process is done because at the time
is was being called otherwise it was too late to enable javascript for that
tab and still have that changed setting affect the page load. Hopefully
nothing breaks because of the signal being emitted before some things like
url() have been updated.