Test for private browsing differently
In a48ea597d021686d94fe06b71e5a28eba9dd0049 we fixed settings in private
QtWebEngine windows.

However, this means we also enable local storage for private windows, which was
disabled in QtWebEngine by default:

https://github.com/qt/qtwebengine/commit/4ef5831a398280cc26cbf359dcbb9878e77ab7fb#diff-44ac7d27348388501944f6a8e2e67d8dR207

It should be safe to enable it, as we get the same behavior as in Chromium, i.e.
a working local storage which entirely lives in RAM.

This also makes those tests work on QtWebKit-NG, presumably because private
browsing for cookies is implemented there.

It also adds a test to at least check whether local storage is isolated from
non-private tabs. I tried writing a test which ensures nothing lands on the hard
disk, but due to QTBUG-52121 this might not happen at all:

https://bugreports.qt.io/browse/QTBUG-52121