Add TOC to installation instructions
This adds a Table of Contents to the installation instructions to
improve the navigation within the document.

I decided to use the command line to configure the TOC because there
were problems with using just using an attribute entry in the document
header.
Specifically the insertion of the `header.asciidoc` into the resulting
HTML file prevented the attribute entry approach from working.

The TOC can now be inserted into any doc file using

    toc::[]