Use upper-case Monospace in test_progress_affecting_statusbar_height
While the test worked again with eb4691adfce64ff9a6b6e60c48489fa9bfd6db20, it
broke again immediately because of 40ee89bddcc863997853cdbf777f9299aa5fe855.

With that fix in, the lower-case monospace in the set value was immediately
replaced by the full list of fonts again. With an upper-case Monospace, this
won't happen.

Fixes #2825, for real this time.