fixup! Greasemonkey: Fix metadata block regex.
The ^ in python regular expressions means "start of the string", not
"start of a line", whoops. This broke the regex for scripts that didn't
have the metadata block at the top of the file, which is definitely
allowed by other implementations.

This does mean we let pass scripts where the UserScript line doesn't
start at column 0 which might be a problem for ... comments describing
how a metadata block works? Well that doesn't sound likely.