Prevent horizontal pan in completion view.
When _resize_columns is called from set_model, the completion view is
not yet shown and the vertical scrollbar is not visible. This caused a
miscalculation in the column widths, resulting in a total width greater
than the window width, which allowed the window to scroll.

This moves the _resize_columns to the showEvent, as well as removing
_maybe_show from setModel, as we now always call set_pattern after
set_model as long as the model is not None.

Resolves #3359.