Greasemonkey: Fetching @required scripts at a init
There was an issue with greasemonkey init happening before a download
manager was available. To avoid having to deal with a qnam directly
scripts that run into that issue will have their processing deferred
until the Application.new_window signal is emitted.

The logic is a bit uglier now that I moved the os.path.exists check to
the top of the function but it is better than trying to cram everything
into one iteration.