Allow download manager to overwrite existing files unprompted.
This is to support the non-interactive use case of setting a
`FileDownloadTarget` and passing auto_remove and not caring if the target
file exists or not.

An alternative to adding the new param to `fetch()` and `set_target()` when it
only works for one target type would be to add it as a parameter on
`FileDownloadTarget` and having set_target pull it out.